Annotated version of Insertion sort
This is the annotated version of insertion sort that can be used only for integer types but can be used for other types with a slight modification. There is also a class in the src folder which can be used for every type but this is more reliable and safe.